rename jartests to tests.jar
