+++ tests/org.eclipse.n4js.ide.tests.jar/.classpath
