+++ tests/org.eclipse.n4js.ide.tests.jar/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.eclipse.n4js.ide.jartests</name>
+	<name>org.eclipse.n4js.ide.tests.jar</name>
 	<comment></comment>
 	<projects>
 	</projects>
