+++ tests/org.eclipse.n4js.ide.tests.jar/.settings/de.loskutov.anyedit.AnyEditTools.prefs
