+++ tests/org.eclipse.n4js.ide.tests.jar/.settings/org.eclipse.emf.ecore.xcore.Xcore.prefs
