+++ tests/org.eclipse.n4js.ide.tests.jar/.settings/org.eclipse.jdt.ui.prefs
