+++ tests/org.eclipse.n4js.ide.tests.jar/build.properties
