+++ tests/org.eclipse.n4js.ide.tests.jar/fixture/P1/package.json
