+++ tests/org.eclipse.n4js.ide.tests.jar/fixture/P2/package.json
