+++ tests/org.eclipse.n4js.ide.tests.jar/fixture/P2/src/k/K.n4js
