+++ tests/org.eclipse.n4js.ide.tests.jar/fixture/PSingle/package.json
