+++ tests/org.eclipse.n4js.ide.tests.jar/fixture/PSingle/src/a/A.n4js
