+++ tests/org.eclipse.n4js.ide.tests.jar/fixture/PSingle/src/b/B.n4js
