+++ tests/org.eclipse.n4js.ide.tests.jar/fixture/PSingle/src/b/b2/B2.n4js
