+++ tests/org.eclipse.n4js.ide.tests.jar/fixture/PSingle/src/c/C.n4js
