+++ tests/org.eclipse.n4js.ide.tests.jar/fixture/PSingleTestNpm/package.json
