+++ tests/org.eclipse.n4js.ide.tests.jar/fixture/PSingleTestNpm/src/Hello.n4js
