+++ tests/org.eclipse.n4js.ide.tests.jar/plugin.properties
