+++ tests/org.eclipse.n4js.ide.tests.jar/pom.xml
@@ -18,8 +18,8 @@ Contributors:
 		<version>0.0.1-SNAPSHOT</version>
 		<relativePath>../pom.xml</relativePath>
 	</parent>
-	<artifactId>org.eclipse.n4js.ide.jartests</artifactId>
-	<name>org.eclipse.n4js.ide.jartests</name>
+	<artifactId>org.eclipse.n4js.ide.tests.jar</artifactId>
+	<name>org.eclipse.n4js.ide.tests.jar</name>
 	<packaging>eclipse-test-plugin</packaging>
 	<description>Test the headless N4js compiler bundled into one jar.</description>
