+++ tests/org.eclipse.n4js.ide.tests.jar/probands/GH-1062/performance-report/package.json
