+++ tests/org.eclipse.n4js.ide.tests.jar/src/org/eclipse/n4js/ide/tests/jar/IncompleteApiImplementationJarTest.java
@@ -8,14 +8,12 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.ide.jartests;
+package org.eclipse.n4js.ide.tests.jar;
 
 import static org.eclipse.n4js.cli.N4jscTestOptions.COMPILE;
-import static org.eclipse.n4js.cli.N4jscTestOptions.HELP;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
-import java.io.File;
-import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
@@ -24,51 +22,47 @@ import org.eclipse.n4js.cli.helper.AbstractCliJarTest;
 import org.eclipse.n4js.cli.helper.CliCompileResult;
 import org.eclipse.n4js.cli.helper.N4CliHelper;
 import org.eclipse.n4js.cli.helper.ProcessResult;
+import org.junit.Ignore;
 import org.junit.Test;
 
 /**
  * IMPORTANT: for info on how to run this test locally, see {@link AbstractCliJarTest}!
  */
-public class SingleFileCompileN4jscJarTest extends AbstractCliJarTest {
+public class IncompleteApiImplementationJarTest extends AbstractCliJarTest {
 
 	/***/
-	public SingleFileCompileN4jscJarTest() {
-		super("fixture");
+	public IncompleteApiImplementationJarTest() {
+		super("IncompleteApiImplementationJarTest");
 	}
 
-	/** Test help command */
+	/**
+	 * Compile & Run Tests for IDE-1510
+	 *
+	 * @throws Exception
+	 *             in Error cases
+	 */
 	@Test
-	public void testHelp() throws Exception {
-		CliCompileResult cliResult = n4jsc(HELP());
-		assertEquals(cliResult.toString(), 0, cliResult.getExitCode());
-	}
+	@Ignore("GH-1291")
+	public void testApiImplStub_CompileAndRunWithNodejsPlugin() throws Exception {
 
-	/** Compile a single project. */
-	@Test
-	public void testSingleFileCompile() {
-		File project = Path.of(TARGET, WORKSPACE_FOLDER, N4CliHelper.PACKAGES, "PSingle").toAbsolutePath().toFile();
-
-		CliCompileResult cliResult = n4jsc(COMPILE(project));
-		assertEquals(cliResult.toString(), 0, cliResult.getExitCode());
-	}
-
-	/** Compile & Run whole project. */
-	@Test
-	public void testCompileAllAndRunWithNodejsPlugin() throws IOException {
 		Path workspace = Path.of(TARGET, WORKSPACE_FOLDER).toAbsolutePath();
 		Path nodeModulesPath = Paths.get(workspace.toString(), N4JSGlobals.NODE_MODULES).toAbsolutePath();
-		Path project = Path.of(workspace.toString(), N4CliHelper.PACKAGES, "P1").toAbsolutePath();
+		Path project = Path.of(workspace.toString(), N4CliHelper.PACKAGES, "one.x.impl").toAbsolutePath();
 
 		N4CliHelper.copyN4jsLibsToLocation(nodeModulesPath, N4JSGlobals.N4JS_RUNTIME);
 
 		CliCompileResult cliResult = n4jsc(COMPILE(project.toFile()));
-		assertEquals(cliResult.toString(), 4, cliResult.getTranspiledFilesCount());
+		assertEquals(cliResult.toString(), 2, cliResult.getTranspiledFilesCount());
 
-		Path fileA = project.resolve("src-gen/A.js");
+		Path fileA = project.resolve("src-gen/AT_IDE-1510_Missing_Method.js");
 
 		ProcessResult nodejsResult = runNodejs(workspace, fileA);
 		assertEquals(nodejsResult.toString(), 0, nodejsResult.getExitCode());
-		assertEquals(nodejsResult.toString(), "Arrghtutut§", nodejsResult.getStdOut());
+		assertTrue(nodejsResult.toString(), nodejsResult.getStdOut()
+				.contains("Hello from Implementation one.x.impl::p.A.n4js !"));
+		assertTrue(nodejsResult.toString(), nodejsResult.getStdOut()
+				.contains("N4ApiNotImplementedError: API for method A.missing_n not implemented yet."));
+		assertTrue(nodejsResult.toString(), nodejsResult.getStdOut()
+				.contains("N4ApiNotImplementedError: API for method A.missing_someIA not implemented yet."));
 	}
-
 }
