+++ tests/org.eclipse.n4js.ide.tests.jar/verifier/verifications-test.xml
