+++ tests/org.eclipse.n4js.ide.tests.jar/xtend-gen/.gitignore
